feat: restructures tests
