+++ packages/sign-client/package.json
@@ -27,7 +27,8 @@
     "test:concurrency": "vitest run --dir test/concurrency",
     "test:xregion": "vitest run --dir test/xregion",
     "test": "npm run test:pre; npm run test:run",
-    "test:ignoreUnhandled": "npm run test:pre; npm run test:run -- --dangerouslyIgnoreUnhandledErrors --no-threads",
+    "test:integration": "vitest run --dir test/sdk/integration",
+    "test:ignoreUnhandled": "npm run test:pre; npm run test:integration -- --dangerouslyIgnoreUnhandledErrors --no-threads",
     "test:canary": "vitest run --dir test/canary",
     "test:lifecycle": "vitest run --dir test/lifecycle",
     "canary": "npm run test:pre; npm run test:canary",
